+++ src/ui/mod.rs
@@ -14,7 +14,6 @@ pub mod palette;
 pub mod pane;
 pub mod perf;
 pub mod presets;
-pub mod profile_editor;
 pub mod settings;
 pub mod sftp;
 pub mod ssh_connect;
